move rgkl crate to rgkl module
