--- crates/rgkl/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "rgkl"
-version = "0.1.0"
-edition = "2021"
-
-
-[dependencies]
-types = {path="../types"}
-
-tokio = { workspace = true, features = ["full"] }
-tokio-util = { workspace = true, features = ["rt"] }
-
-tonic = { workspace = true, features = ["gzip"] }
-
-chrono = { workspace = true}
-prost = { workspace = true}
-prost-types = { workspace = true}
-tempfile = { workspace = true }
-notify = { workspace = true, features = ["crossbeam-channel", "macos_kqueue"] }
-thiserror = { workspace = true }
-
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["env-filter", "fmt", "registry", "std"] }
-tracing-test = { workspace = true }
-
-grep = "0.3.2"
-grep-cli = "0.1.11"
-grep-matcher = "0.1.7"
-grep-printer = "0.2.2"
-grep-regex = "0.1.13"
-grep-searcher = "0.1.14"
-memchr = "2.7.5"
-rand = "0.9.2"
-regex = "1.11.1"
-rstest = "0.26.1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-signal-hook = "0.3.18"
-termcolor = "1.4.1"
